refactor(diff): split chunked.rs into module with format leaf
Maintainability guardrail: chunked.rs grew past the 600 production-line
target. Move the wire format (constants, serialize/deserialize,
ChunkedPatchData, bitset helpers, read helpers) into chunked/format.rs;
chunked/mod.rs keeps the public diff/apply API and options. Test suites
follow their subject: format structure/compat tests in format.rs,
roundtrip tests in mod.rs.
